+++ webapp/src/routes/api/download-microservices/shared.ts
@@ -1,9 +1,6 @@
 import type { TemplatesResponse } from '$lib/pocketbase/types';
 import type AdmZip from 'adm-zip';
-
-//
-
-export const DEFAULT_LOCALE = 'en-US';
+import { getZipRootFolder } from './utils/zip';
 
 //
 
@@ -25,6 +22,16 @@ export function getFoldersToDelete(microserviceToKeep: MicroserviceFolder): stri
 
 //
 
+export function getCredentialCustomCodePath(
+	zip: AdmZip,
+	microservice: MicroserviceFolder,
+	credential_type_name: string
+) {
+	const CUSTOM_CODE_FOLDER = 'custom_code';
+	const zipRoot = getZipRootFolder(zip);
+	return `${zipRoot}${microservice}/${CUSTOM_CODE_FOLDER}/${credential_type_name}`;
+}
+
 export function addCustomCode(
 	zip: AdmZip,
 	microservice: MicroserviceFolder,
@@ -32,8 +39,7 @@ export function addCustomCode(
 	template: TemplatesResponse
 ) {
 	const { zencode_data, zencode_script } = template;
-	const CUSTOM_CODE_FOLDER = 'custom_code';
-	const basePath = `/${microservice}/${CUSTOM_CODE_FOLDER}/${credential_type_name}`;
+	const basePath = getCredentialCustomCodePath(zip, microservice, credential_type_name);
 	zip.addFile(`${basePath}.zen`, Buffer.from(zencode_script));
 	zip.addFile(`${basePath}.data.json`, Buffer.from(zencode_data));
 }